refactor: type-checking cleanup and minor fixes
- Add type annotations and `# type: ignore` comments for untyped
  pygame_gui calls (kill, select, show, hide, rebuild, etc.)
- Add return type hints to closures and setUp methods
- Use `cast()` in ServiceLocator.get() and GameDriver helpers
- Wrap numeric returns in explicit `float()` for type safety
- Rename shadowed loop variables (button→btn, info→sq_info)
- Annotate dynamic dicts with `dict[str, Any]`
- Fix AnimationDefinition import path in animation_helpers
- Remove unused `Any` import from lazy_loader
- Add `# type: ignore[import-untyped]` for simpleeval
- Remove duplicate README headers and stale pytest lines
- Update .agent/rules to add Ty preference over Mypy
- Re-enable skipped social context tests
- Fix driver.render fallback to always pass alpha arg

+++ src/yukkuri_game/benchmarks/benchmark.py
@@ -86,7 +86,7 @@ class BenchmarkRunner:
         """
         if psutil:
             process = psutil.Process()
-            return process.memory_info().rss / 1024 / 1024
+            return float(process.memory_info().rss / 1024 / 1024)
         return 0.0
 
     def _get_process_cpu(self) -> float:
@@ -98,7 +98,7 @@ class BenchmarkRunner:
         """
         if psutil:
             process = psutil.Process()
-            return process.cpu_percent(interval=None)
+            return float(process.cpu_percent(interval=None))
         return 0.0
 
     def _get_system_info(self) -> dict[str, str]:
